+++ src/main/java/wooteco/subway/ui/ControllerAdvice.java
@@ -3,7 +3,7 @@ package wooteco.subway.ui;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.dto.ErrorMessageResponse;
+import wooteco.subway.dto.response.ErrorMessageResponse;
 
 @RestControllerAdvice
 public class ControllerAdvice {
