+++ src/test/java/wooteco/subway/acceptance/LineAcceptanceTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.dto.LineResponse;
+import wooteco.subway.dto.response.LineResponse;
 
 class LineAcceptanceTest extends AcceptanceTest {
